main: fix possible infinite loop
Fix an infinite loop that can happen if we don't handle a WebsocketError
reading from next. We should rather propagate the error to the outer
loop and let the outer loop trigger a reconnect.

While at it also handle a None read from the WebSocketConnection by
triggering a reconnect as well.